Prevent accidental rm -rf /*
Adds a safer way to do the final rm -rf of the scratch dir.

+++ expose.sh
@@ -917,7 +917,7 @@ do
 		cd "$topdir"
 	fi
 	
-	rm -rf "$scratchdir"/*
+	rm -rf "${scratchdir:?}/"*
 done
 
 # copy resources to _site
